requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
audioread==3.0.1
bayesian-optimization==1.5.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
dash==3.2.0
decorator==5.2.1
einops==0.8.1
essentia==2.1b6.dev1389
essentia-tensorflow==2.1b6.dev1389
faiss-cpu==1.12.0
filelock==3.13.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2024.6.1
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.7
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.9.2
lightning==2.3.0
lightning-utilities==0.15.2
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mido==1.3.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
narwhals==1.43.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.22.0
packaging==24.2
pandas==2.3.0
piano-transcription-inference==0.0.6
pillow==11.0.0
platformdirs==4.3.8
plotly==6.1.2
pooch==1.8.2
pretty_midi==0.2.10
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pynndescent==0.5.13
pyparsing==3.2.3
pysnmp==7.1.21
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
resampy==0.4.3
retrying==1.4.2
rpds-py==0.25.1
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.34.1
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
streamlit==1.46.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchlibrosa==0.1.0
torchmetrics==1.8.1
torchvision==0.21.0+cpu
tornado==6.5.1
tqdm==4.67.1
transformers==4.55.0
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wandb==0.21.1
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
